✨ <feat> add files for feedback form.

+++ src/main/java/projetb2/overlook_hotel/controller/api/BookingApiController.java
@@ -29,7 +29,7 @@ public class BookingApiController {
      * with an error message.
      */
     @PostMapping("/accept")
-    public RedirectView acceptBooking(@RequestParam("bookingId") Long bookingId,
+    public RedirectView acceptBooking(@RequestParam("bookingId") Integer bookingId,
             RedirectAttributes redirectAttributes) {
         boolean result = bookingService.acceptBooking(bookingId);
         redirectAttributes.addFlashAttribute("message", result ? "Booking accepted" : "Could not accept booking");
@@ -44,7 +44,7 @@ public class BookingApiController {
      * with an error message.
      */
     @PostMapping("/decline")
-    public RedirectView declineBooking(@RequestParam("bookingId") Long bookingId,
+    public RedirectView declineBooking(@RequestParam("bookingId") Integer bookingId,
             RedirectAttributes redirectAttributes) {
         boolean result = bookingService.declineBooking(bookingId);
         redirectAttributes.addFlashAttribute("message", result ? "Booking declined" : "Could not decline booking");
@@ -58,7 +58,7 @@ public class BookingApiController {
      * page with a success or error message.
      */
     @PostMapping("/edit")
-    public RedirectView editBooking(@RequestParam Long bookingId,
+    public RedirectView editBooking(@RequestParam Integer bookingId,
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate arrivingDate,
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate departureDate,
             @RequestParam(required = false) BookingStatus bookingStatus,
